Refactor: move `AccountService` back to deployment.
The methods in `AccountService` will need to take parameters which are specific to the deployment subsystem, i.e., study deployment information. Therefore, this cannot be in common. Likely, each different subsystem that has to do account operations needs to implement those in a separate `AccountService`.
