--- carp.common/src/commonTest/kotlin/dk/cachet/carp/common/users/InMemoryAccountServiceTest.kt
@@ -1,10 +1,0 @@
-package dk.cachet.carp.common.users
-
-
-/**
- * Tests for [InMemoryAccountService].
- */
-class InMemoryAccountServiceTest : AccountServiceTest()
-{
-    override fun createService(): AccountService = InMemoryAccountService()
-}
