+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/domain/users/AccountService.kt
@@ -1,4 +1,7 @@
-package dk.cachet.carp.common.users
+package dk.cachet.carp.deployment.domain.users
+
+import dk.cachet.carp.common.users.Account
+import dk.cachet.carp.common.users.AccountIdentity
 
 
 /**
@@ -7,15 +10,15 @@ package dk.cachet.carp.common.users
 interface AccountService
 {
     /**
-     * Create a new account for the person identified by [identity] to participate in a study.
-     * The invitation and account details should be delivered, or made available, to the person managing the [identity].
+     * Create a new account identified by [identity] to participate in a study.
+     * The invitation and account details should be delivered, or made available, to the user managing the [identity].
      *
      * @throws IllegalArgumentException when an account with a matching [AccountIdentity] already exists.
      */
     suspend fun inviteNewAccount( identity: AccountIdentity ): Account
 
     /**
-     * Deliver an invitation to participate in a study, or make it available, to the person managing [identity].
+     * Deliver an invitation to participate in a study, or make it available, to the user managing [identity].
      *
      * @throws IllegalArgumentException when no account with a matching [identity] exists.
      */
