+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/domain/InMemoryStudyRepository.kt
@@ -22,7 +22,7 @@ class InMemoryStudyRepository : StudyRepository
 
 
 /**
- * Tests whether the [InMemoryUserRepository] stub is implemented correctly.
+ * Tests whether the [InMemoryStudyRepository] stub is implemented correctly.
  */
 class InMemoryStudyRepositoryTest : StudyRepositoryTest
 {
